[lld] Add support for delay load helper.
This is different than MS link.exe. Instead of going through an emulator
for delay loading, add a separated chunk type that takes care of it.

Since CRT has no support for modifying auxiliary delay IAT, this
moves aux IAT to .data section and modifies it directly in the chunk.